HIVE-28801Iceberg: Refactor HMS table parameter setting to be able to reuse
